http-proxy-agent: don't mix in the `proxy` opts to the endpoint opts
This is done on the `https-proxy-agent` module so that TLS options can
be used on the destination endpoint. However, on this module, there is
no TLS destination endpoint option, so mixing in the `proxy` opts here
makes no sense.

Also it makes the produced absolute URL incorrect if the hostname doesn't
match the proxy server's (i.e. most of the time...). So this fixes
TooTallNate/superagent-proxy#3 as well :)